bbhunt/requirements.txt
click>=8.0.0
prompt_toolkit>=3.0.0
PyYAML>=6.0
psutil>=5.9.0
requests>=2.27.1
redis>=4.3.0
fastapi>=0.95.0
uvicorn>=0.21.0
python-dotenv>=1.0.0
dnspython>=2.3.0
